+++ src/pat/slides/composed-injection-test.html
@@ -1,32 +1,42 @@
-<!DOCTYPE HTML>
+<!DOCTYPE html>
 <html lang="en-US">
-<head>
-	<title>Customer cases — Informaat</title>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=1274, user-scalable=no">
-	<link rel="stylesheet" href="slideshow.css" media="all">
-        <script data-main="../main" src="../../src/bower_components/requirejs/require.js" type="text/javascript" charset="utf-8"></script>
-</head>
-<body class="list">
-	<div class="pat-slides">		
-	    <a href="standalone-test.html" class="pat-inject" data-pat-inject="source: #third-slide::element; target: self::element; trigger: autoload">Run</a>
-		<div class="slide" id="first-slide">
-			<div class="slide-content">
-				<section>
-					<header>
-						<h2>My beautiful slide</h2>
-					</header>
-					<p>
-						This slide is not injected but all the previous ones were.
-					</p>
-				</section>
-			</div>
-                        <footer>
-				<p>
-				    Some notes.
-				</p>
-			</footer>
-		</div>
-	</div>
-</body>
+  <head>
+    <title>Customer cases — Informaat</title>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=1274, user-scalable=no" />
+    <link rel="stylesheet" href="slideshow.css" media="all" />
+    <script
+      data-main="../main"
+      src="../../src/bower_components/requirejs/require.js"
+      type="text/javascript"
+      charset="utf-8"
+    ></script>
+  </head>
+  <body class="list">
+    <div class="pat-slides">
+      <a
+        href="standalone-test.html"
+        class="pat-inject"
+        data-pat-inject="source: #third-slide::element; target: self::element; trigger: autoload"
+        >Run</a
+      >
+      <div class="slide" id="first-slide">
+        <div class="slide-content">
+          <section>
+            <header>
+              <h2>My beautiful slide</h2>
+            </header>
+            <p>
+              This slide is not injected but all the previous ones were.
+            </p>
+          </section>
+        </div>
+        <footer>
+          <p>
+            Some notes.
+          </p>
+        </footer>
+      </div>
+    </div>
+  </body>
 </html>
