+++ src/pat/slides/index.html
@@ -1,19 +1,29 @@
-<!DOCTYPE HTML>
+<!DOCTYPE html>
 <html lang="en-US">
   <head>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-    <title>Slideshows </title>
-    <link rel="stylesheet" href="/style/common.css" type="text/css">
-         <script src="/bundle.js" type="text/javascript" charset="utf-8"></script>
- </head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+    <title>Slideshows</title>
+    <link rel="stylesheet" href="/style/common.css" type="text/css" />
+    <script src="/bundle-vendors.js" type="text/javascript" charset="utf-8"></script>
+    <script src="/bundle.js" type="text/javascript" charset="utf-8"></script>
+  </head>
   <body>
-        <ul>
-            <li><a href="standalone-test.html">Standalone slideshow page</a></li>
-            <li><a href="standalone-test.html">Page with injected slides</a></li>
-            <li><a href="standalone-test.html">Page with a composition of embedded and injected slides</a></li>
-        </ul>
-        <ul>
-            <li><a href="standalone-test.html?slides=first-slide,third-slide">Standalone slideshow page with filters for the first and the third slide.</a></li>
-        </ul>
+    <ul>
+      <li><a href="standalone-test.html">Standalone slideshow page</a></li>
+      <li><a href="standalone-test.html">Page with injected slides</a></li>
+      <li>
+        <a href="standalone-test.html"
+          >Page with a composition of embedded and injected slides</a
+        >
+      </li>
+    </ul>
+    <ul>
+      <li>
+        <a href="standalone-test.html?slides=first-slide,third-slide"
+          >Standalone slideshow page with filters for the first and the third
+          slide.</a
+        >
+      </li>
+    </ul>
   </body>
 </html>
