+++ src/pat/slides/injection-test.html
@@ -1,13 +1,23 @@
-<!DOCTYPE HTML>
+<!DOCTYPE html>
 <html lang="en-US">
-<head>
-	<title>Customer cases — Informaat</title>
-	<meta charset="UTF-8"/>
-	<meta name="viewport" content="width=1274, user-scalable=no"/>
-	<link rel="stylesheet" href="slideshow.css" media="all"/>
-	<script data-main="../main" src="../../src/bower_components/requirejs/require.js" type="text/javascript" charset="utf-8"></script>
-</head>
-<body class="list">
-	<a href="standalone-test.html?slides=first-slide,last-slide" class="pat-inject" data-pat-inject="target: self::element; trigger: autoload">Run</a>
-</body>
+  <head>
+    <title>Customer cases — Informaat</title>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=1274, user-scalable=no" />
+    <link rel="stylesheet" href="slideshow.css" media="all" />
+    <script
+      data-main="../main"
+      src="../../src/bower_components/requirejs/require.js"
+      type="text/javascript"
+      charset="utf-8"
+    ></script>
+  </head>
+  <body class="list">
+    <a
+      href="standalone-test.html?slides=first-slide,last-slide"
+      class="pat-inject"
+      data-pat-inject="target: self::element; trigger: autoload"
+      >Run</a
+    >
+  </body>
 </html>
