+++ src/pat/slides/notes.md
@@ -1,9 +1,11 @@
 # Presentation notes
 
 ## Introduction
+
 …
 
 ## Dutch government
+
 The Binnenhof (Dutch pronunciation: [ˈbɪnənɦɔf], inner court), is a complex of buildings in The Hague. It has been the location of meetings of the Staten-Generaal, the Dutch parliament, since 1446, and has been the centre of Dutch politics for many centuries.
 
 The grounds on which the Binnenhof now stands were purchased by Count Floris IV of Holland in 1229, where he built his mansion, next to the little lake that has been called Hofvijver or 'Court Pond' since the 13th century. More buildings were constructed around the court, several of which are well known in their own right, such as the Ridderzaal (Great hall; literally Knight's Hall), where the queen holds her annual speech at Prinsjesdag. One of the towers, simply known as het Torentje ('the Little Tower'; directly next to the Mauritshuis museum) has been the office of the Prime Minister of the Netherlands since 1982.
@@ -12,16 +14,18 @@ This 'Inner Court' is studded with monumental old buildings testifying of eight 
 
 A large modern building at the south side of the Binnenhof since 1992 houses the House of Representatives, the lower but more important of the Dutch democratically elected Houses of parliament.
 
-* * *
+---
 
 Source: Wikipedia
 
-
 ## Dialogues
-Dialogues are the brick and mortar of UX design. Customers interact with organisation through touchpoints. These touchpoints used to be quite physical like shops, help desks or intermediates. Today, more and more of these touchpoints are digital. This made customers also users. The moment of interaction between a user and an organisation — during which the user typically performs a task — is called a dialogue. 
+
+Dialogues are the brick and mortar of UX design. Customers interact with organisation through touchpoints. These touchpoints used to be quite physical like shops, help desks or intermediates. Today, more and more of these touchpoints are digital. This made customers also users. The moment of interaction between a user and an organisation — during which the user typically performs a task — is called a dialogue.
 
 ## Animations
-Like to use animations on your slides? Eat your heart out with animate.css to quickly add fancy animation to your slides. Add your own animations to top it off like with this rebellious pearl. 
+
+Like to use animations on your slides? Eat your heart out with animate.css to quickly add fancy animation to your slides. Add your own animations to top it off like with this rebellious pearl.
 
 ## Cards
+
 Play your cards well!
