+++ src/pat/slides/style/animate.css/README.md
@@ -1,5 +1,5 @@
 #Animate.css
-*Just-add-water CSS animation*
+_Just-add-water CSS animation_
 
 `animate.css` is a bunch of cool, fun, and cross-browser animations for you to use in your projects. Great for emphasis, home pages, sliders, and general just-add-water-awesomeness.
 
@@ -9,29 +9,29 @@ To use animate.css in your website, simply drop the stylesheet into your documen
 You can do a whole bunch of other stuff with animate.css when you combine it with jQuery or add your own CSS rules. Dynamically add animations using jQuery with ease:
 
 ```javascript
-$('#yourElement').addClass('animated bounceOutLeft');
+$("#yourElement").addClass("animated bounceOutLeft");
 ```
 
 You can change the duration of your animations, add a delay or change the number of times that it plays!
 
 ```css
 #yourElement {
-	-vendor-animation-duration: 3s;
-	-vendor-animation-delay: 2s;
-	-vendor-animation-iteration-count: infinite;
+  -vendor-animation-duration: 3s;
+  -vendor-animation-delay: 2s;
+  -vendor-animation-iteration-count: infinite;
 }
 ```
 
-*Note: be sure to replace "vendor" in the CSS with the applicable vendor prefixes (webkit, moz, ms, o)*
+_Note: be sure to replace "vendor" in the CSS with the applicable vendor prefixes (webkit, moz, ms, o)_
 
-*Note: Safari in Mountion Lion (OS 10.8) has a display glitch with the Flippers. They may not appear at all until the animation is completed, or the page may have other artifacting. One fix is to add overflow: hidden to the parent div.*
+_Note: Safari in Mountion Lion (OS 10.8) has a display glitch with the Flippers. They may not appear at all until the animation is completed, or the page may have other artifacting. One fix is to add overflow: hidden to the parent div._
 
 ##License
 Animate.css is licensed under the &#9786; license. (http://licence.visualidiot.com/)
 
 ##Learn more
 You can learn more about animate.css over at http://daneden.me/animate
-You can also get in touch via email (dan.eden@me.com) or twitter (@_dte) if you need any help or have any issues.
+You can also get in touch via email (dan.eden@me.com) or twitter (@\_dte) if you need any help or have any issues.
 
 ##Cheat Sheet
