+++ src/pat/slides/style/fontello/css/fontello-codes.css
@@ -1,56 +1,161 @@
 @charset "UTF-8";
 
-
-.icon-home:before { content: '\2302'; } /* '⌂' */
-.icon-star:before { content: '\2605'; } /* '★' */
-.icon-star-empty:before { content: '\2606'; } /* '☆' */
-.icon-list:before { content: '\2630'; } /* '☰' */
-.icon-list-large:before { content: '\e807'; } /* '' */
-.icon-heart-empty:before { content: '\e800'; } /* '' */
-.icon-heart:before { content: '\e801'; } /* '' */
-.icon-tiles:before { content: '\268f'; } /* '⚏' */
-.icon-grid-large:before { content: '\e80b'; } /* '' */
-.icon-flag:before { content: '\2691'; } /* '⚑' */
-.icon-cog:before { content: '\e804'; } /* '' */
-.icon-cog-many:before { content: '\26ef'; } /* '⛯' */
-.icon-mail:before { content: '\2709'; } /* '✉' */
-.icon-edit:before { content: '\270d'; } /* '✍' */
-.icon-pencil:before { content: '\270e'; } /* '✎' */
-.icon-ok-circle-1:before { content: '\2714'; } /* '✔' */
-.icon-cancel-circle:before { content: '\e803'; } /* '' */
-.icon-plus-circle:before { content: '\e802'; } /* '' */
-.icon-minus-circle:before { content: '\2796'; } /* '➖' */
-.icon-ccw:before { content: '\27f2'; } /* '⟲' */
-.icon-user-add:before { content: '\e700'; } /* '' */
-.icon-ok-circle:before { content: '\e702'; } /* '' */
-.icon-cancel-circle-open:before { content: '\e703'; } /* '' */
-.icon-help-circle:before { content: '\e704'; } /* '' */
-.icon-info-circle:before { content: '\e705'; } /* '' */
-.icon-eye:before { content: '\e70a'; } /* '' */
-.icon-tag:before { content: '\e70c'; } /* '' */
-.icon-tags:before { content: '\e70d'; } /* '' */
-.icon-upload-cloud:before { content: '\e711'; } /* '' */
-.icon-export:before { content: '\e80a'; } /* '' */
-.icon-print:before { content: '\e716'; } /* '' */
-.icon-comment:before { content: '\e806'; } /* '' */
-.icon-chat:before { content: '\e805'; } /* '' */
-.icon-vcard:before { content: '\e722'; } /* '' */
-.icon-trash:before { content: '\e729'; } /* '' */
-.icon-rss:before { content: '\e73a'; } /* '' */
-.icon-resize-full:before { content: '\e744'; } /* '' */
-.icon-resize-small:before { content: '\e746'; } /* '' */
-.icon-network:before { content: '\e776'; } /* '' */
-.icon-picture:before { content: '🌄'; } /* '\1f304' */
-.icon-thumbs-up:before { content: '👍'; } /* '\1f44d' */
-.icon-user:before { content: '👤'; } /* '\1f464' */
-.icon-users:before { content: '👥'; } /* '\1f465' */
-.icon-folder:before { content: '📁'; } /* '\1f4c1' */
-.icon-pin:before { content: '📌'; } /* '\1f4cc' */
-.icon-attach:before { content: '📎'; } /* '\1f4ce' */
-.icon-phone:before { content: '📞'; } /* '\1f4de' */
-.icon-upload:before { content: '\e808'; } /* '' */
-.icon-download:before { content: '\e809'; } /* '' */
-.icon-search:before { content: '🔍'; } /* '\1f50d' */
-.icon-lock:before { content: '🔒'; } /* '\1f512' */
-.icon-lock-open:before { content: '🔓'; } /* '\1f513' */
-.icon-link:before { content: '🔗'; } /* '\1f517' */
+.icon-home:before {
+  content: "\2302";
+} /* '⌂' */
+.icon-star:before {
+  content: "\2605";
+} /* '★' */
+.icon-star-empty:before {
+  content: "\2606";
+} /* '☆' */
+.icon-list:before {
+  content: "\2630";
+} /* '☰' */
+.icon-list-large:before {
+  content: "\e807";
+} /* '' */
+.icon-heart-empty:before {
+  content: "\e800";
+} /* '' */
+.icon-heart:before {
+  content: "\e801";
+} /* '' */
+.icon-tiles:before {
+  content: "\268f";
+} /* '⚏' */
+.icon-grid-large:before {
+  content: "\e80b";
+} /* '' */
+.icon-flag:before {
+  content: "\2691";
+} /* '⚑' */
+.icon-cog:before {
+  content: "\e804";
+} /* '' */
+.icon-cog-many:before {
+  content: "\26ef";
+} /* '⛯' */
+.icon-mail:before {
+  content: "\2709";
+} /* '✉' */
+.icon-edit:before {
+  content: "\270d";
+} /* '✍' */
+.icon-pencil:before {
+  content: "\270e";
+} /* '✎' */
+.icon-ok-circle-1:before {
+  content: "\2714";
+} /* '✔' */
+.icon-cancel-circle:before {
+  content: "\e803";
+} /* '' */
+.icon-plus-circle:before {
+  content: "\e802";
+} /* '' */
+.icon-minus-circle:before {
+  content: "\2796";
+} /* '➖' */
+.icon-ccw:before {
+  content: "\27f2";
+} /* '⟲' */
+.icon-user-add:before {
+  content: "\e700";
+} /* '' */
+.icon-ok-circle:before {
+  content: "\e702";
+} /* '' */
+.icon-cancel-circle-open:before {
+  content: "\e703";
+} /* '' */
+.icon-help-circle:before {
+  content: "\e704";
+} /* '' */
+.icon-info-circle:before {
+  content: "\e705";
+} /* '' */
+.icon-eye:before {
+  content: "\e70a";
+} /* '' */
+.icon-tag:before {
+  content: "\e70c";
+} /* '' */
+.icon-tags:before {
+  content: "\e70d";
+} /* '' */
+.icon-upload-cloud:before {
+  content: "\e711";
+} /* '' */
+.icon-export:before {
+  content: "\e80a";
+} /* '' */
+.icon-print:before {
+  content: "\e716";
+} /* '' */
+.icon-comment:before {
+  content: "\e806";
+} /* '' */
+.icon-chat:before {
+  content: "\e805";
+} /* '' */
+.icon-vcard:before {
+  content: "\e722";
+} /* '' */
+.icon-trash:before {
+  content: "\e729";
+} /* '' */
+.icon-rss:before {
+  content: "\e73a";
+} /* '' */
+.icon-resize-full:before {
+  content: "\e744";
+} /* '' */
+.icon-resize-small:before {
+  content: "\e746";
+} /* '' */
+.icon-network:before {
+  content: "\e776";
+} /* '' */
+.icon-picture:before {
+  content: "🌄";
+} /* '\1f304' */
+.icon-thumbs-up:before {
+  content: "👍";
+} /* '\1f44d' */
+.icon-user:before {
+  content: "👤";
+} /* '\1f464' */
+.icon-users:before {
+  content: "👥";
+} /* '\1f465' */
+.icon-folder:before {
+  content: "📁";
+} /* '\1f4c1' */
+.icon-pin:before {
+  content: "📌";
+} /* '\1f4cc' */
+.icon-attach:before {
+  content: "📎";
+} /* '\1f4ce' */
+.icon-phone:before {
+  content: "📞";
+} /* '\1f4de' */
+.icon-upload:before {
+  content: "\e808";
+} /* '' */
+.icon-download:before {
+  content: "\e809";
+} /* '' */
+.icon-search:before {
+  content: "🔍";
+} /* '\1f50d' */
+.icon-lock:before {
+  content: "🔒";
+} /* '\1f512' */
+.icon-lock-open:before {
+  content: "🔓";
+} /* '\1f513' */
+.icon-link:before {
+  content: "🔗";
+} /* '\1f517' */
